+++ go.mod
@@ -1,3 +1,3 @@
-module git.litelot.us/litelotus/simulacra
+module github.com/LotusLabsSoftworks/simulacra-server
 
 go 1.24.2
